+++ application/IFrameProcessor.hpp
@@ -1,0 +1,8 @@
+#include <opencv2/core/core.hpp>
+
+class IFrameProcessor {
+ public:
+    virtual cv::Mat process(cv::Mat input) =0;
+};
+
+IFrameProcessor* createToGrayscaleProcessor();
