+++ application/IFrameSource.hpp
@@ -1,7 +1,6 @@
 # pragma once
 
 #include <opencv2/core/core.hpp>
-#include <opencv2/highgui/highgui.hpp>
 
 class IFrameSource
 {
@@ -9,27 +8,4 @@ class IFrameSource
      virtual cv::Mat getFrame() = 0;
 };
 
-class ImageFrameSource: public IFrameSource
-{
- public:
-    ImageFrameSource(std::string path_to_image) {
-        image_ = cv::imread(path_to_image);
-    }
-
-    cv::Mat getFrame() {
-        // We allow to read image only once
-        if (!image_.empty())
-        {
-            cv::Mat clone = image_.clone();
-            image_ = cv::Mat();
-            return clone;
-        }
-        else
-        {
-            return image_;
-        }
-    }
-
- private:
-     cv::Mat image_;
-};
+IFrameSource* createImageFrameSource(std::string path_to_image);
